delete checks for authorized user successfully

+++ Server/controllers/cover-letter.controller.js
@@ -21,9 +21,12 @@ const CoverLetterController = {
     const id = req.params.id;
     try {
       let coverLetter = await CoverLetter.findById(id);
-      console.log(coverLetter);
       if (!coverLetter) {
         return res.status(404).json({ message: "Cover-Letter not found" });
+      }
+      console.log(coverLetter);
+      if (coverLetter.userID.toString() !== req.user._id.toString()) {
+        return res.status(401).json({ message: "This is not your Cover-Letter" });
       }
       await coverLetter.deleteOne();
       return res.status(201).json({ message: "Cover-Letter is successfully deleted", coverLetter });
